// src/app/models/movie.ts
export class Movie{
    id: string;
    originalTitle: string;
    overview: string;
    popularity: number;
    releaseDate: Date;
    posterPath: string;
    Title: string;
    video: boolean;
    voteAverage: number;
    voteCount: number;
    Genre = [];
    Language: string;
    budget: string;
    revenue: string;
    runtime: string;
    status: number;
    profet: string;
    trailer: any;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
    title: string;
    Country: string;
    totalResults: number = 0;
    Path: string;
}